refactor : 컴포넌트 분리 및 멘토님 리뷰 수정
