+++ src/apis/cartAPI.ts
@@ -1,17 +1,15 @@
 import instance from './axios';
 import { PostOrderItem, PostClient, PostSubscriber } from 'types/Orders';
-const cartAPI = {
-	postRoomToCart(
+export const postRoomToCart = (
 		roomTypeId: number,
 		checkInDate: string,
 		checkOutDate: string,
-	) {
+	)  => {
 		return instance.post('carts', {
 			roomTypeId: roomTypeId,
 			checkinDate: checkInDate,
 			checkoutDate: checkOutDate,
 		});
-	},
 };
 
 // 장바구니 개별 삭제
@@ -47,4 +45,4 @@ export const getOrderCheck = async (orderId: number) => {
 	const res = await instance.get(`orders/${orderId}`);
 	return res.data;
 };
-export default cartAPI;
+
